+++ CHANGELOG.md
@@ -37,6 +37,7 @@ Each section shall contain a list of action items of the following format: `<bri
 - Semantic markup for profiles, profile-options, requirements, gateways, use-cases to support rich artefact generation ([#463, part 2](https://github.com/IHE/DEV.SDPi/issues/463))
 - OID summary table for profiles, actors, transactions, content modules, use cases, requirements ([#449](https://github.com/IHE/DEV.SDPi/issues/449))
 - Implementation conformity statement tables ([#463, part 2](https://github.com/IHE/DEV.SDPi/issues/463))
+- Added processing support to include and support implementation conformance statements from external standards ([#463, part 3](https://github.com/IHE/DEV.SDPi/issues/463))
 - Relations extension to model inter-MDS relationships ([#48, part1](https://github.com/IHE/DEV.SDPi/issues/48))
 
 ### Changes